reorganizes and updates novelty report scripts
