+++ projects/out_of_sample/scripts/nd-report/config-examples/yearlong-run/offline_prognostic_report_config.yaml
@@ -1,0 +1,139 @@
+models:
+  - name: "ocsvm-4/79-0.4"
+    model_url: "gs://vcm-ml-experiments/claytons/2022-07-18/ocsvm-gamma-trial-1/trained_models/ocsvm-large"
+    cutoff: 0.4
+  - name: "ocsvm-4/79-0.2"
+    model_url: "gs://vcm-ml-experiments/claytons/2022-07-18/ocsvm-gamma-trial-1/trained_models/ocsvm-large"
+    cutoff: 0.2
+  - name: "ocsvm-4/79-0"
+    model_url: "gs://vcm-ml-experiments/claytons/2022-07-18/ocsvm-gamma-trial-1/trained_models/ocsvm-large"
+    cutoff: 0
+  - name: "ocsvm-4/79--0.5"
+    model_url: "gs://vcm-ml-experiments/claytons/2022-07-18/ocsvm-gamma-trial-1/trained_models/ocsvm-large"
+    cutoff: -0.5
+  - name: "ocsvm-4/79--1"
+    model_url: "gs://vcm-ml-experiments/claytons/2022-07-18/ocsvm-gamma-trial-1/trained_models/ocsvm-large"
+    cutoff: -1
+  - name: "ocsvm-4/79--1.5"
+    model_url: "gs://vcm-ml-experiments/claytons/2022-07-18/ocsvm-gamma-trial-1/trained_models/ocsvm-large"
+    cutoff: -1.5
+  - name: "ocsvm-tq-2/79-0.2"
+    model_url: "gs://vcm-ml-experiments/claytons/2022-08-09/tq-combined-nd-trial-1/trained_models/ocsvm-large-gamma"
+    cutoff: 0.2
+  - name: "ocsvm-tq-2/79-0"
+    model_url: "gs://vcm-ml-experiments/claytons/2022-08-09/tq-combined-nd-trial-1/trained_models/ocsvm-large-gamma"
+    cutoff: 0
+  - name: "ocsvm-tq-4/79-0.2"
+    model_url: "gs://vcm-ml-experiments/claytons/2022-08-09/tq-combined-nd-trial-1/trained_models/ocsvm-larger-gamma"
+    cutoff: 0.2
+  - name: "ocsvm-tq-4/79-0"
+    model_url: "gs://vcm-ml-experiments/claytons/2022-08-09/tq-combined-nd-trial-1/trained_models/ocsvm-larger-gamma"
+    cutoff: 0
+  - name: "ocsvm-1/79-4"
+    model_url: "gs://vcm-ml-experiments/claytons/2022-07-15/ocsvm-default-trial-3/trained_models/ocsvm"
+    cutoff: 4
+  - name: "ocsvm-1/79-2"
+    model_url: "gs://vcm-ml-experiments/claytons/2022-07-15/ocsvm-default-trial-3/trained_models/ocsvm"
+    cutoff: 2
+  - name: "ocsvm-1/79-0"
+    model_url: "gs://vcm-ml-experiments/claytons/2022-07-15/ocsvm-default-trial-3/trained_models/ocsvm"
+    cutoff: 0
+  - name: "ocsvm-1/79--2"
+    model_url: "gs://vcm-ml-experiments/claytons/2022-07-15/ocsvm-default-trial-3/trained_models/ocsvm"
+    cutoff: -2
+  - name: "ocsvm-1/79--4"
+    model_url: "gs://vcm-ml-experiments/claytons/2022-07-15/ocsvm-default-trial-3/trained_models/ocsvm"
+    cutoff: -4
+  - name: "ocsvm-1/79--6"
+    model_url: "gs://vcm-ml-experiments/claytons/2022-07-15/ocsvm-default-trial-3/trained_models/ocsvm"
+    cutoff: -6
+  - name: "ocsvm-1/4/79-5"
+    model_url: "gs://vcm-ml-experiments/claytons/2022-07-18/ocsvm-gamma-trial-1/trained_models/ocsvm-small"
+    cutoff: 5
+  - name: "ocsvm-1/4/79-0"
+    model_url: "gs://vcm-ml-experiments/claytons/2022-07-18/ocsvm-gamma-trial-1/trained_models/ocsvm-small"
+    cutoff: 0
+  - name: "ocsvm-1/4/79--5"
+    model_url: "gs://vcm-ml-experiments/claytons/2022-07-18/ocsvm-gamma-trial-1/trained_models/ocsvm-small"
+    cutoff: -5
+  - name: "ocsvm-1/4/79--10"
+    model_url: "gs://vcm-ml-experiments/claytons/2022-07-18/ocsvm-gamma-trial-1/trained_models/ocsvm-small"
+    cutoff: -10
+  - name: "ocsvm-1/4/79--15"
+    model_url: "gs://vcm-ml-experiments/claytons/2022-07-18/ocsvm-gamma-trial-1/trained_models/ocsvm-small"
+    cutoff: -15
+  - name: "ocsvm-1/4/79--20"
+    model_url: "gs://vcm-ml-experiments/claytons/2022-07-18/ocsvm-gamma-trial-1/trained_models/ocsvm-small"
+    cutoff: -20
+  - name: "minmax-0.2"
+    model_url: "gs://vcm-ml-experiments/claytons/2022-07-13/oos-test/novelty"
+    cutoff: 0.2
+  - name: "minmax-0.1"
+    model_url: "gs://vcm-ml-experiments/claytons/2022-07-13/oos-test/novelty"
+    cutoff: 0.1
+  - name: "minmax-0"
+    model_url: "gs://vcm-ml-experiments/claytons/2022-07-13/oos-test/novelty"
+    cutoff: 0
+  - name: "minmax--0.03"
+    model_url: "gs://vcm-ml-experiments/claytons/2022-07-13/oos-test/novelty"
+    cutoff: -0.03
+  - name: "minmax--0.06"
+    model_url: "gs://vcm-ml-experiments/claytons/2022-07-13/oos-test/novelty"
+    cutoff: -0.06
+  - name: "minmax--0.1"
+    model_url: "gs://vcm-ml-experiments/claytons/2022-07-13/oos-test/novelty"
+    cutoff: -0.1
+  - name: "ocsvm-tq-4/79-0.2"
+    model_url: "gs://vcm-ml-experiments/claytons/2022-07-18/ocsvm-gamma-trial-1/trained_models/ocsvm-large"
+    cutoff: 0.2
+  - name: "ocsvm-tq-4/79-0"
+    model_url: "gs://vcm-ml-experiments/claytons/2022-07-18/ocsvm-gamma-trial-1/trained_models/ocsvm-large"
+    cutoff: 0
+  - name: "ocsvm-tq-4/79--0.5"
+    model_url: "gs://vcm-ml-experiments/claytons/2022-07-18/ocsvm-gamma-trial-1/trained_models/ocsvm-large"
+    cutoff: -0.5
+  - name: "ocsvm-tq-4/79--1"
+    model_url: "gs://vcm-ml-experiments/claytons/2022-07-18/ocsvm-gamma-trial-1/trained_models/ocsvm-large"
+    cutoff: -1
+  - name: "ocsvm-tq-1/79-2"
+    model_url: "gs://vcm-ml-experiments/claytons/2022-07-15/ocsvm-default-trial-3/trained_models/ocsvm"
+    cutoff: 2
+  - name: "ocsvm-tq-1/79-0"
+    model_url: "gs://vcm-ml-experiments/claytons/2022-07-15/ocsvm-default-trial-3/trained_models/ocsvm"
+    cutoff: 0
+  - name: "ocsvm-tq-1/79--2"
+    model_url: "gs://vcm-ml-experiments/claytons/2022-07-15/ocsvm-default-trial-3/trained_models/ocsvm"
+    cutoff: -2
+  - name: "ocsvm-tq-1/79--4"
+    model_url: "gs://vcm-ml-experiments/claytons/2022-07-15/ocsvm-default-trial-3/trained_models/ocsvm"
+    cutoff: -4
+  - name: "ocsvm-tq-1/4/79-5"
+    model_url: "gs://vcm-ml-experiments/claytons/2022-07-18/ocsvm-gamma-trial-1/trained_models/ocsvm-small"
+    cutoff: 5
+  - name: "ocsvm-tq-1/4/79-0"
+    model_url: "gs://vcm-ml-experiments/claytons/2022-07-18/ocsvm-gamma-trial-1/trained_models/ocsvm-small"
+    cutoff: 0
+  - name: "ocsvm-tq-1/4/79--5"
+    model_url: "gs://vcm-ml-experiments/claytons/2022-07-18/ocsvm-gamma-trial-1/trained_models/ocsvm-small"
+    cutoff: -5
+  - name: "ocsvm-tq-1/4/79--10"
+    model_url: "gs://vcm-ml-experiments/claytons/2022-07-18/ocsvm-gamma-trial-1/trained_models/ocsvm-small"
+    cutoff: -10
+  - name: "minmax-tq-0.2"
+    model_url: "gs://vcm-ml-experiments/claytons/2022-07-13/oos-test/novelty"
+    cutoff: 0.2
+  - name: "minmax-tq-0.1"
+    model_url: "gs://vcm-ml-experiments/claytons/2022-07-13/oos-test/novelty"
+    cutoff: 0.1
+  - name: "minmax-tq-0"
+    model_url: "gs://vcm-ml-experiments/claytons/2022-07-13/oos-test/novelty"
+    cutoff: 0
+  - name: "minmax-tq--0.03"
+    model_url: "gs://vcm-ml-experiments/claytons/2022-07-13/oos-test/novelty"
+    cutoff: -0.03
+run_dataset:
+  name: "prognostic"
+  url: "gs://vcm-ml-experiments/claytons/2022-08-10/oos-wind-vanilla-trial2/fv3gfs_run"
+report_url: "gs://vcm-ml-public/claytons/oos_analysis/baseline-cutoffs/"
+has_cutoff_plots: false
+append_random_id_to_url: true
